Refactor: Extract Farmhand component logic into useFarmhand hook
